requirements.txt
flask
streamlit
pandas
numpy
scikit-learn
joblib
requests
tldextract
python-whois
beautifulsoup4
